+++ src/ingest/network.rs
@@ -54,10 +54,10 @@ impl Display for Conn {
 }
 
 impl ResponseRangeData for Conn {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let conn_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let conn_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &conn_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &conn_csv.as_bytes())))
     }
 }
 
@@ -211,10 +211,10 @@ impl Display for Dns {
 }
 
 impl ResponseRangeData for Dns {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let dns_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let dns_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &dns_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &dns_csv.as_bytes())))
     }
 }
 
@@ -292,10 +292,10 @@ impl Display for Http {
 }
 
 impl ResponseRangeData for Http {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let http_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let http_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &http_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &http_csv.as_bytes())))
     }
 }
 
@@ -327,10 +327,10 @@ impl Display for Rdp {
 }
 
 impl ResponseRangeData for Rdp {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let rdp_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let rdp_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &rdp_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &rdp_csv.as_bytes())))
     }
 }
 
@@ -374,10 +374,10 @@ impl Display for Smtp {
 }
 
 impl ResponseRangeData for Smtp {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let smtp_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let smtp_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &smtp_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &smtp_csv.as_bytes())))
     }
 }
 
@@ -417,10 +417,10 @@ impl Display for Ntlm {
 }
 
 impl ResponseRangeData for Ntlm {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let ntlm_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let ntlm_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &ntlm_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &ntlm_csv.as_bytes())))
     }
 }
 
@@ -468,10 +468,10 @@ impl Display for Kerberos {
 }
 
 impl ResponseRangeData for Kerberos {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let kerberos_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let kerberos_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &kerberos_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &kerberos_csv.as_bytes())))
     }
 }
 
@@ -527,10 +527,10 @@ impl Display for Ssh {
 }
 
 impl ResponseRangeData for Ssh {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let ssh_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let ssh_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &ssh_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &ssh_csv.as_bytes())))
     }
 }
 
@@ -568,10 +568,10 @@ impl Display for DceRpc {
 }
 
 impl ResponseRangeData for DceRpc {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let dce_rpc_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let dce_rpc_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &dce_rpc_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &dce_rpc_csv.as_bytes())))
     }
 }
 
@@ -625,10 +625,10 @@ impl Display for Ftp {
 }
 
 impl ResponseRangeData for Ftp {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let ftp_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let ftp_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &ftp_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &ftp_csv.as_bytes())))
     }
 }
 
@@ -670,10 +670,10 @@ impl Display for Mqtt {
 }
 
 impl ResponseRangeData for Mqtt {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let mqtt_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let mqtt_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &mqtt_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &mqtt_csv.as_bytes())))
     }
 }
 
@@ -717,10 +717,10 @@ impl Display for Ldap {
 }
 
 impl ResponseRangeData for Ldap {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let ldap_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let ldap_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &ldap_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &ldap_csv.as_bytes())))
     }
 }
 
@@ -792,10 +792,10 @@ impl Display for Tls {
 }
 
 impl ResponseRangeData for Tls {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let tls_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let tls_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &tls_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &tls_csv.as_bytes())))
     }
 }
 
@@ -848,10 +848,10 @@ impl Display for Smb {
 }
 
 impl ResponseRangeData for Smb {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let smb_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let smb_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &smb_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &smb_csv.as_bytes())))
     }
 }
 
@@ -885,10 +885,10 @@ impl Display for Nfs {
 }
 
 impl ResponseRangeData for Nfs {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let nfs_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let nfs_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &nfs_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &nfs_csv.as_bytes())))
     }
 }
 
@@ -940,10 +940,10 @@ impl Display for Bootp {
 }
 
 impl ResponseRangeData for Bootp {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let bootp_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let bootp_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &bootp_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &bootp_csv.as_bytes())))
     }
 }
 
@@ -1009,9 +1009,9 @@ impl Display for Dhcp {
 }
 
 impl ResponseRangeData for Dhcp {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let dhcp_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let dhcp_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &dhcp_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &dhcp_csv.as_bytes())))
     }
 }