+++ src/ingest/sysmon.rs
@@ -70,10 +70,10 @@ impl Display for ProcessCreate {
 }
 
 impl ResponseRangeData for ProcessCreate {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let process_create_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let process_create_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &process_create_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &process_create_csv.as_bytes())))
     }
 }
 
@@ -109,10 +109,10 @@ impl Display for FileCreationTimeChanged {
 }
 
 impl ResponseRangeData for FileCreationTimeChanged {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let file_create_time_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let file_create_time_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &file_create_time_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &file_create_time_csv.as_bytes())))
     }
 }
 
@@ -166,10 +166,10 @@ impl Display for NetworkConnection {
 }
 
 impl ResponseRangeData for NetworkConnection {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let network_connect_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let network_connect_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &network_connect_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &network_connect_csv.as_bytes())))
     }
 }
 
@@ -199,12 +199,12 @@ impl Display for ProcessTerminated {
 }
 
 impl ResponseRangeData for ProcessTerminated {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let process_terminate_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let process_terminate_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
         bincode::serialize(&Some((
             timestamp,
-            source,
+            sensor,
             &process_terminate_csv.as_bytes(),
         )))
     }
@@ -255,10 +255,10 @@ impl Display for ImageLoaded {
 }
 
 impl ResponseRangeData for ImageLoaded {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let image_load_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let image_load_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &image_load_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &image_load_csv.as_bytes())))
     }
 }
 
@@ -292,10 +292,10 @@ impl Display for FileCreate {
 }
 
 impl ResponseRangeData for FileCreate {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let file_create_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let file_create_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &file_create_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &file_create_csv.as_bytes())))
     }
 }
 
@@ -331,13 +331,13 @@ impl Display for RegistryValueSet {
 }
 
 impl ResponseRangeData for RegistryValueSet {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
         let registry_value_set_csv =
-            format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+            format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
         bincode::serialize(&Some((
             timestamp,
-            source,
+            sensor,
             &registry_value_set_csv.as_bytes(),
         )))
     }
@@ -375,13 +375,13 @@ impl Display for RegistryKeyValueRename {
 }
 
 impl ResponseRangeData for RegistryKeyValueRename {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
         let registry_key_rename_csv =
-            format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+            format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
         bincode::serialize(&Some((
             timestamp,
-            source,
+            sensor,
             &registry_key_rename_csv.as_bytes(),
         )))
     }
@@ -421,13 +421,13 @@ impl Display for FileCreateStreamHash {
 }
 
 impl ResponseRangeData for FileCreateStreamHash {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
         let file_create_stream_hash_csv =
-            format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+            format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
         bincode::serialize(&Some((
             timestamp,
-            source,
+            sensor,
             &file_create_stream_hash_csv.as_bytes(),
         )))
     }
@@ -463,10 +463,10 @@ impl Display for PipeEvent {
 }
 
 impl ResponseRangeData for PipeEvent {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let pipe_event_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let pipe_event_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &pipe_event_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &pipe_event_csv.as_bytes())))
     }
 }
 
@@ -502,10 +502,10 @@ impl Display for DnsEvent {
 }
 
 impl ResponseRangeData for DnsEvent {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let dns_event_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let dns_event_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &dns_event_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &dns_event_csv.as_bytes())))
     }
 }
 
@@ -543,10 +543,10 @@ impl Display for FileDelete {
 }
 
 impl ResponseRangeData for FileDelete {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let file_delete_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let file_delete_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &file_delete_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &file_delete_csv.as_bytes())))
     }
 }
 
@@ -578,10 +578,10 @@ impl Display for ProcessTampering {
 }
 
 impl ResponseRangeData for ProcessTampering {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
-        let process_tamper_csv = format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
+        let process_tamper_csv = format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
-        bincode::serialize(&Some((timestamp, source, &process_tamper_csv.as_bytes())))
+        bincode::serialize(&Some((timestamp, sensor, &process_tamper_csv.as_bytes())))
     }
 }
 
@@ -617,13 +617,13 @@ impl Display for FileDeleteDetected {
 }
 
 impl ResponseRangeData for FileDeleteDetected {
-    fn response_data(&self, timestamp: i64, source: &str) -> Result<Vec<u8>, bincode::Error> {
+    fn response_data(&self, timestamp: i64, sensor: &str) -> Result<Vec<u8>, bincode::Error> {
         let file_delete_detected_csv =
-            format!("{}\t{source}\t{self}", convert_time_format(timestamp));
+            format!("{}\t{sensor}\t{self}", convert_time_format(timestamp));
 
         bincode::serialize(&Some((
             timestamp,
-            source,
+            sensor,
             &file_delete_detected_csv.as_bytes(),
         )))
     }